:truck: move lib modules to src
into : src/api, src/utils, and src/types. also adds a apiVersion to the sanity client config
