+++ src/components/MealPreview.tsx
@@ -3,10 +3,10 @@ import { default as NextLink } from 'next/link'
 import { Box, Flex, Grid, Text, Card, Heading, Link } from 'theme-ui'
 import { CardProps } from '@theme-ui/components'
 import { Clock } from 'phosphor-react'
-import minutesToHours from '../../lib/minutesToHours'
+import minutesToHours from '../utils/minutesToHours'
 import * as listStyles from '../constants/styles/stripedList'
 import * as nestedStyles from '../constants/styles/nested'
-import { Meal } from '../../lib/types'
+import { Meal } from '../types/api'
 import Inline from './Inline'
 
 type level = 'h2' | 'h3' | 'h4' | 'h5' | 'h5'
