+++ src/utils/pluralize.ts
